fix(ui): keep the Notra component file component-only
Move the SVG string helper out so Fast Refresh can preserve state.
